# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "O-Ushers API",
    "description": "The **O-Ushers** platform connects event organizers with professional ushers (talents).\n\n## Roles\n| Role | Description |\n|------|-------------|\n| **Usher** | A talent who browses events, applies, and manages their profile |\n| **Organizer** | Posts events, manages applicants, marks attendance & reviews |\n| **Organizer Member** | Uses the organizer workspace without event or staff CRUD access |\n| **Organizer Supervisor** | Uses the organizer workspace and manages event operations |\n| **Admin** | Platform-level management: users, events, verification |\n\n## Authentication\nBrowser clients authenticate with secure HTTP-only cookies. API clients may also send the\nshort-lived access token returned by `POST /auth/login` as a Bearer token.\n\n## Pagination\nPaginated endpoints accept `page` and `size` query parameters and return:\n```json\n{ \"data\": [], \"currentPage\": 1, \"totalPages\": 5, \"totalCount\": 42 }\n```\n",
    "version": "1.0.0",
    "contact": {
      "email": "ooushers@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://o-ushers.vercel.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/verify/{token}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify email address",
        "description": "Called when the user clicks the verification link sent to their email\nafter signup. Returns an HTML success/failure page.\n",
        "security": [],
        "operationId": "verifyEmail",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "JWT token embedded in the verification email link."
          }
        ],
        "responses": {
          "200": {
            "description": "HTML page — verification succeeded"
          },
          "401": {
            "description": "HTML page — token invalid or expired"
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List current user's notifications",
        "operationId": "getNotifications",
        "responses": {
          "200": {
            "description": "Up to 100 newest notifications"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Clear current user's notifications",
        "operationId": "clearNotifications",
        "responses": {
          "200": {
            "description": "Notifications cleared"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT obtained from `POST /auth/login`. Browser clients normally\nrely on the secure HTTP-only session cookies instead.\n"
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid Bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Something went wrong"
          }
        }
      },
...
    }
  }
}